+++ src/categories/dtos/input/pagination-options-category.input.ts
