+++ src/users/resolvers/users.resolver.ts
@@ -4,7 +4,7 @@ import { Roles } from 'src/auth/decorators/roles.decorator';
 import { GqlJwtGuard } from 'src/auth/guards/gql-jwt.guard';
 import { GqlRolesGuard } from 'src/auth/guards/gql-roles.guard';
 import { Role } from 'src/auth/roles.enum';
-import { PaginationOptionsUserInput } from '../dtos/input/pagination-user.input';
+import { PaginationOptionsUserInput } from '../dtos/input/pagination-options-user.input';
 import { UpdateUserInput } from '../dtos/input/update-user.input';
 import { PaginatedUser } from '../models/paginated-user.model';
 import { User } from '../models/user.model';
